test(remote-lease): split contract+ibc tests into per-topic sub-modules
Both files were approaching 1000 LoC; co-locating tests by topic makes
navigation faster and keeps unrelated helpers from leaking into shared
scope.

contract/tests/{mod,instantiate,migrate,query,execute_admin,execute_outbound,scenarios}.rs
  - Split by entry point. mod.rs holds shared constants + helpers used
    by two or more sub-modules (deps, deps_with_lease, instantiate_msg,
    sender, store_open_channel, store_closing_channel, sample_open_lease_params,
    query_config, query_channel).
  - Sub-module-local helpers stay private:
    - deps_with_failing_code_info in instantiate.rs
    - migrate_msg in migrate.rs
    - assert_send_packet + expected_timeout + sample_swap/transfer_out + open_lease_execute in execute_outbound.rs
    - open_channel_via_admin + drive_open_ack + drive_open_lease + handshake_channel + local_endpoint + ack_msg_with + assert_callback_to in scenarios.rs

ibc/tests/{mod,handshake,packets}.rs
  - Split by IBC phase. mod.rs holds shared constants + deps_with_config + channel().
  - handshake.rs owns all ibc_channel_* tests (open + connect + close) and the
    persist_existing_* helpers used only by them.
  - packets.rs owns ibc_packet_* tests (receive + ack + timeout + wire-shape
    pin + the ADR 0001 fixtures) and the packet-construction helpers
    (envelope_with_close_lease, encode_envelope, outbound_packet, ack_msg,
    timeout_msg, assert_dispatched_callback).

No test added, removed, or renamed. 72/72 pass; cargo fmt + clippy clean.

Largest module after the split is packets.rs at 394 LoC. The split is one
indivisible refactor: intermediate states would not compile (mod.rs
referencing sub-modules that did not yet exist, or sub-modules with no
parent mod.rs declaration).
